# pkg/api/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cloud-Barista cm-centipede — Multi-cloud filesystem, object storage, and DBMS data migration subsystem.",
        "title": "cm-centipede API",
        "contact": {
            "name": "Cloud-Barista Community",
            "url": "https://github.com/cloud-barista/cm-centipede"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8085",
    "basePath": "/",
    "paths": {
...
        "/centipede/migration/{migrationId}/validation": {
            "get": {
                "description": "Returns the current validation status, summary message, and per-item detail list for the given migration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "validation"
                ],
                "summary": "Get validation result for a migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Migration ID",
                        "name": "migrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_ValidationResultResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            },
            "post": {
                "description": "Validates that the destination data matches the source after a completed migration. Returns 400 if migration is not completed, 409 if validation is already running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "validation"
                ],
                "summary": "Start validation for a completed migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Migration ID",
                        "name": "migrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_SimpleMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/centipede/plans/target": {
            "post": {
                "description": "Validates source/destination connections, checks DBMS compatibility and version, and returns a TargetDataMigrationModel ready for CreateMigration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plan"
                ],
                "summary": "Analyse target migration plan",
                "parameters": [
                    {
                        "description": "Target plan request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.TargetPlanReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_dmdl_target-model_TargetDataMigrationModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleDBRef": {
            "type": "object",
            "required": [
                "nsId",
                "password",
                "rdbmsId"
            ],
            "properties": {
                "nsId": {
                    "type": "string"
                },
                "password": {
                    "description": "AES-encrypted on store",
                    "type": "string"
                },
                "rdbmsId": {
                    "type": "string"
                },
                "tlsCAPem": {
                    "description": "TLSCAPEM is the PEM root that verify-ca and verify-full check against.\nEmpty means the system trust store, which carries only publicly trusted\nCAs — and managed RDBMS services commonly sign with their own, so those two\nmodes usually need the CSP's root supplied here.",
                    "type": "string"
                },
                "tlsMode": {
                    "description": "TLSMode selects transport security; the empty value is \"disable\". It means\nthe same thing here as on DBConnConfig — see the table in transx-ex's\ndbmsx/driver/base/tls.go.",
                    "type": "string",
                    "enum": [
                        "disable",
                        "prefer",
                        "require",
                        "verify-ca",
                        "verify-full"
                    ]
                },
                "username": {
                    "description": "defaults to the looked-up AdminUsername",
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleOSRef": {
            "type": "object",
            "required": [
                "nsId",
                "osId"
            ],
            "properties": {
                "nsId": {
                    "type": "string"
                },
                "osId": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleSSHRef": {
            "type": "object",
            "required": [
                "infraId",
                "nodeId",
                "nsId"
            ],
            "properties": {
                "infraId": {
                    "type": "string"
                },
                "nodeId": {
                    "type": "string"
                },
                "nsId": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef": {
            "type": "object",
            "required": [
                "source"
            ],
            "properties": {
                "beetleDb": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleDBRef"
                },
                "beetleObjectStorage": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleOSRef"
                },
                "beetleSsh": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleSSHRef"
                },
                "db": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBConnConfig"
                },
                "honeybee": {
                    "description": "Reference types",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.HoneybeeRef"
                        }
                    ]
                },
                "minio": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.MinioConnConfig"
                },
                "source": {
                    "enum": [
                        "honeybee",
                        "beetleSsh",
                        "beetleObjectStorage",
                        "beetleDb",
                        "ssh",
                        "minio",
                        "db"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionSource"
                        }
                    ]
                },
                "ssh": {
                    "description": "Inline types",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.SSHConnConfig"
                        }
                    ]
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionSource": {
            "type": "string",
            "enum": [
                "honeybee",
                "beetleSsh",
                "beetleObjectStorage",
                "beetleDb",
                "ssh",
                "minio",
                "db"
            ],
            "x-enum-varnames": [
                "ConnectionSourceHoneybee",
                "ConnectionSourceBeetleSSH",
                "ConnectionSourceBeetleObjectStorage",
                "ConnectionSourceBeetleDB",
                "ConnectionSourceSSH",
                "ConnectionSourceMinio",
                "ConnectionSourceDB"
            ]
        },
...
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilter": {
            "type": "object",
            "properties": {
                "databases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilterEntry"
                    }
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilterEntry": {
            "type": "object",
            "required": [
                "targetMapping"
            ],
            "properties": {
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilterRule"
                    }
                },
                "targetMapping": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.TargetMapping"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.FileSystemFilter": {
            "type": "object",
            "required": [
                "targetMapping"
            ],
            "properties": {
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.PathFilterRule"
                    }
                },
                "targetMapping": {
                    "description": "selects the single folder to migrate",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.TargetMapping"
                        }
                    ]
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.HoneybeeRef": {
            "type": "object",
            "required": [
                "connectionId"
            ],
            "properties": {
                "connectionId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.ObjectStorageFilter": {
            "type": "object",
            "required": [
                "targetMapping"
            ],
            "properties": {
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.PathFilterRule"
                    }
                },
                "targetMapping": {
                    "description": "selects the single bucket to migrate",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.TargetMapping"
                        }
                    ]
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.PathFilterRule": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "include | exclude",
                    "type": "string"
                },
                "op": {
                    "description": "type=size (\u003e \u003e= \u003c \u003c= ==)",
                    "type": "string"
                },
                "pattern": {
                    "description": "type=glob",
                    "type": "string"
                },
                "type": {
                    "description": "glob | size",
                    "type": "string"
                },
                "value": {
                    "description": "type=size (bytes)",
                    "type": "integer"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_dmdl_common-model.SSHConnConfig": {
            "type": "object",
            "required": [
                "host",
                "username"
            ],
            "properties": {
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "description": "defaults to 22",
                    "type": "integer"
                },
                "privateKey": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDataMigrationModel": {
            "type": "object",
            "properties": {
                "sourceDataMigrationModel": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDataMigrationProperty"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationModel": {
            "type": "object",
            "required": [
                "targetDataMigrationModel"
            ],
            "properties": {
                "targetDataMigrationModel": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationProperty"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_dmdl_target-model_TargetDataMigrationModel": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationModel"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_SimpleMessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.SimpleMessageResponse"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_ValidationResultResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ValidationResultResponse"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.SimpleMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.TargetPlanReq": {
            "type": "object",
            "required": [
                "dstConnection",
                "source"
            ],
            "properties": {
                "dbmsFilter": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilter"
                },
                "dstConnection": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef"
                },
                "fileSystemFilter": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.FileSystemFilter"
                },
                "fileSystemStrategy": {
                    "description": "FileSystemStrategy is the transfer strategy applied to filesystem (SSH→SSH)\ntasks. Empty is normalised to \"relay\" (safer default: avoids agent-forward\nsetup requirements). It has no effect on object-storage or DBMS tasks, where\nrelay is always used / not applicable.",
                    "type": "string",
                    "enum": [
                        "auto",
                        "direct",
                        "relay"
                    ]
                },
                "objectStorageFilter": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ObjectStorageFilter"
                },
                "source": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDataMigrationModel"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ValidationDetailItem": {
            "type": "object",
            "properties": {
                "itemPath": {
                    "description": "ItemPath: file path | object key | DBMS target\n(\"{database}/{table}\", \"{database}/{table}.{column}\", \"{database}/{kind}:{name}\").",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "Status: \"passed\" | \"failed\" | \"warning\". A warning is reported but does not\nfail the validation — a character-set difference can be a managed service\nimposing its own default.",
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ValidationResultResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "list of failed items",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ValidationDetailItem"
                    }
                },
                "validationMessage": {
                    "type": "string"
                },
                "validationStatus": {
                    "description": "\"passed\" | \"failed\" | \"not_run\" | \"running\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}